Switch from getopts to manual argument reading
Also check if variables are set to ensure correct function.

+++ task1/task1b.sh
@@ -1,22 +1,31 @@
 #!/bin/sh
 
-while getopts ":h" opt; do
-    case $opt in
-        h)
-            printf "Description:\n"
-            printf "\tThis script creates a \"tmp\" with 5 directories in it in the home directory, and creates two text files within each subdirectory. The text files contain the first name and last name of the user which are read as environment variables from \$FIRST_NAME and \$LAST_NAME.\n\n"
-            printf "Usage:\n"
-            printf "\ttask1b.sh [-h]\n"
-            printf "Options:\n"
-            printf "\t-h: Displays this help message and exits.\n"
-            exit 0
+usage() {
+        printf "Description:\n"
+        printf "\tThis script creates a \"tmp\" with 5 directories in it in the home directory, and creates two text files within each subdirectory. The text files contain the first name and last name of the user which are read as environment variables from \$FIRST_NAME and \$LAST_NAME.\n\n"
+        printf "Usage:\n"
+        printf "\ttask1b.sh [-h]\n"
+        printf "Options:\n"
+        printf "\t-h: Displays this help message and exits.\n"
+        exit 0
+    }
+
+while [ $# -gt 0 ]; do
+    case $1 in
+        -h)
+            usage
             ;;
-        \?)
-            printf "Invalid option: -$OPTARG\n" >&2
+        *)
+            echo "Invalid option: $1" >&2
             exit 1
             ;;
     esac
 done
+
+if [ -z "$FIRST_NAME" ] && [ -z "$LAST_NAME" ]; then
+    echo "Error: FIRST_NAME and LAST_NAME environment variables are not set."
+    exit 1
+fi
 
 rm -r ~/tmp
